+++ mayan/apps/converter/backends/python.py
@@ -8,6 +8,10 @@ from PIL import Image
 import PyPDF2
 import sh
 import yaml
+try:
+    from yaml import CSafeLoader as SafeLoader
+except ImportError:
+    from yaml import SafeLoader
 
 from django.utils.encoding import force_text
 from django.utils.translation import ugettext_lazy as _
@@ -25,7 +29,9 @@ from ..literals import (
 
 try:
     pdftoppm = sh.Command(
-        yaml.load(setting_graphics_backend_config.value).get(
+        yaml.load(
+            stream=setting_graphics_backend_config.value, Loader=SafeLoader
+        ).get(
             'pdftoppm_path', DEFAULT_PDFTOPPM_PATH
         )
     )
@@ -33,13 +39,17 @@ except sh.CommandNotFound:
     pdftoppm = None
 else:
     pdftoppm_format = '-{}'.format(
-        yaml.load(setting_graphics_backend_config.value).get(
+        yaml.load(
+            stream=setting_graphics_backend_config.value, Loader=SafeLoader
+        ).get(
             'pdftoppm_format', DEFAULT_PDFTOPPM_FORMAT
         )
     )
 
     pdftoppm_dpi = format(
-        yaml.load(setting_graphics_backend_config.value).get(
+        yaml.load(
+            stream=setting_graphics_backend_config.value, Loader=SafeLoader
+        ).get(
             'pdftoppm_dpi', DEFAULT_PDFTOPPM_DPI
         )
     )
@@ -48,7 +58,9 @@ else:
 
 try:
     pdfinfo = sh.Command(
-        yaml.load(setting_graphics_backend_config.value).get(
+        yaml.load(
+            stream=setting_graphics_backend_config.value, Loader=SafeLoader
+        ).get(
             'pdfinfo_path', DEFAULT_PDFINFO_PATH
         )
     )