+++ mayan/apps/converter/classes.py
@@ -9,6 +9,11 @@ from PIL import Image
 import sh
 import yaml
 
+try:
+    from yaml import CSafeLoader as SafeLoader
+except ImportError:
+    from yaml import SafeLoader
+
 from django.utils.translation import ugettext_lazy as _
 
 from mayan.apps.common.settings import setting_temporary_directory
@@ -26,7 +31,9 @@ logger = logging.getLogger(__name__)
 
 try:
     LIBREOFFICE = sh.Command(
-        yaml.load(setting_graphics_backend_config.value).get(
+        yaml.load(
+            stream=setting_graphics_backend_config.value, Loader=SafeLoader
+        ).get(
             'libreoffice_path', DEFAULT_LIBREOFFICE_PATH
         )
     ).bake('--headless', '--convert-to', 'pdf:writer_pdf_Export')
@@ -181,7 +188,7 @@ class ConverterBase(object):
 
     def get_page(self, output_format=None, as_base64=False):
         output_format = output_format or yaml.load(
-            setting_graphics_backend_config.value
+            stream=setting_graphics_backend_config.value, Loader=SafeLoader
         ).get(
             'pillow_format', DEFAULT_PILLOW_FORMAT
         )