+++ mayan/apps/converter/forms.py
@@ -2,6 +2,11 @@ from __future__ import unicode_literals
 
 import yaml
 
+try:
+    from yaml import CSafeLoader as SafeLoader
+except ImportError:
+    from yaml import SafeLoader
+
 from django import forms
 from django.core.exceptions import ValidationError
 from django.utils.translation import ugettext_lazy as _
@@ -16,7 +21,7 @@ class TransformationForm(forms.ModelForm):
 
     def clean(self):
         try:
-            yaml.safe_load(self.cleaned_data['arguments'])
+            yaml.load(stream=self.cleaned_data['arguments'], Loader=SafeLoader)
         except yaml.YAMLError:
             raise ValidationError(
                 _(