+++ mayan/apps/converter/validators.py
@@ -2,6 +2,11 @@ from __future__ import unicode_literals
 
 import yaml
 
+try:
+    from yaml import CSafeLoader as SafeLoader
+except ImportError:
+    from yaml import SafeLoader
+
 from django.core.exceptions import ValidationError
 from django.utils.deconstruct import deconstructible
 from django.utils.translation import ugettext_lazy as _
@@ -15,7 +20,7 @@ class YAMLValidator(object):
     def __call__(self, value):
         value = value.strip()
         try:
-            yaml.safe_load(value)
+            yaml.load(stream=value, Loader=SafeLoader)
         except yaml.error.YAMLError:
             raise ValidationError(
                 _('Enter a valid YAML value.'),