+++ mayan/apps/document_signatures/storages.py
@@ -2,6 +2,11 @@ from __future__ import unicode_literals
 
 import yaml
 
+try:
+    from yaml import CSafeLoader as SafeLoader
+except ImportError:
+    from yaml import SafeLoader
+
 from django.utils.module_loading import import_string
 
 from .settings import (
@@ -11,7 +16,8 @@ from .settings import (
 storage_detachedsignature = import_string(
     dotted_path=setting_storage_backend.value
 )(
-    **yaml.safe_load(
-        setting_storage_backend_arguments.value or '{}'
+    **yaml.load(
+        stream=setting_storage_backend_arguments.value or '{}',
+        Loader=SafeLoader
     )
 )