+++ mayan/apps/smart_settings/forms.py
@@ -2,6 +2,11 @@ from __future__ import unicode_literals
 
 import yaml
 
+try:
+    from yaml import CSafeLoader as SafeLoader
+except ImportError:
+    from yaml import SafeLoader
+
 from django import forms
 from django.core.exceptions import ValidationError
 from django.utils.translation import ugettext_lazy as _
@@ -33,7 +38,7 @@ class SettingForm(forms.Form):
                 )
 
         try:
-            yaml.safe_load(self.cleaned_data['value'])
+            yaml.load(stream=self.cleaned_data['value'], Loader=SafeLoader)
         except yaml.YAMLError:
             raise ValidationError(
                 _(