+++ mayan/apps/sources/storages.py
@@ -1,6 +1,10 @@
 from __future__ import unicode_literals
 
 import yaml
+try:
+    from yaml import CSafeLoader as SafeLoader
+except ImportError:
+    from yaml import SafeLoader
 
 from django.utils.module_loading import import_string
 
@@ -12,7 +16,8 @@ from .settings import (
 storage_staging_file_image_cache = import_string(
     dotted_path=setting_staging_file_image_cache_storage.value
 )(
-    **yaml.safe_load(
-        setting_staging_file_image_cache_storage_arguments.value or '{}'
+    **yaml.load(
+        stream=setting_staging_file_image_cache_storage_arguments.value or '{}',
+        Loader=SafeLoader
     )
 )