+++ HISTORY.rst
@@ -23,6 +23,9 @@
 * Remove development URLs from main URL file.
 * Move API documentation generation from the root URLs module
   to the API app's URLs module. 
+* Update Pillow to version 6.0.0
+* Update PyYAML to version 5.1. Update use of safe_load and
+  safe_dump to load and dump using the SafeLoader.
 
 3.1.11 (2019-04-XX)
 ===================